14.12.2021

+++ Entity/Concrete/Campaign.cs
@@ -12,6 +12,6 @@ namespace Entity.Concrete
         public int Id { get; set; }
         public string Name { get; set; }
         public DateTime FinishDate { get; set; }
-        public double DiscountPercentage { get; set; }
+        public int DiscountPercentage { get; set; }
     }
 }
